src/core/patterns/arithmetic_dynamics.py
"""Arithmetic Dynamics Layer.

This module implements the arithmetic dynamics layer that captures
deep structural patterns in computational motives. It bridges:

- Arithmetic Height Theory
- Dynamical Systems
- Motivic Integration
- L-functions and Modular Forms

The core idea is that computational patterns have an inherent
arithmetic structure that can be understood through dynamical
systems over adelic spaces.
"""

from typing import Dict, List, Tuple, Any, Union

import numpy as np
import torch
from torch import nn
from torchtyping import TensorType, patch_typeguard
from typeguard import typechecked
from torchinfo import summary

patch_typeguard()  # Enable runtime shape checking


class ArithmeticDynamics(nn.Module):
    """Arithmetic dynamics computation with quantum corrections."""

    def __init__(
        self,
        hidden_dim: int,
        motive_rank: int,
        num_primes: int = 8,  # Number of prime bases for adelic structure
        height_dim: int = 4,  # Dimension of height space
        quantum_weight: float = 0.1,  # Weight for quantum corrections
    ):
        super().__init__()

        self.hidden_dim = hidden_dim
        self.motive_rank = motive_rank
        self.num_primes = num_primes
        self.height_dim = height_dim
        self.quantum_weight = quantum_weight

        # Initialize prime bases (first num_primes primes)
        self.register_buffer(
            "prime_bases",
            torch.tensor(
                [2, 3, 5, 7, 11, 13, 17, 19][:num_primes], dtype=torch.float32
            ),
        )

        # Arithmetic structure
        self.height_map = nn.Linear(hidden_dim, height_dim)

        # Flow computation
        self.flow = nn.Linear(height_dim, height_dim)

        # L-function computation - adjusted for batched inputs
        self.l_function = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim // 2),
            nn.SiLU(),
            nn.Linear(hidden_dim // 2, motive_rank)
        )

        # Adelic projection
        self.adelic_proj = nn.Linear(hidden_dim, num_primes * motive_rank)

        # Output projection layers
        self.measure_proj = nn.Linear(height_dim, 2)  # Project to 2D measure space
        self.output_proj = nn.Linear(2, hidden_dim)  # Project back to input space

        # Dynamical system parameters
        self.coupling = nn.Parameter(torch.randn(num_primes, motive_rank))

        # Quantum correction networks
        self.quantum_height = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim),
            nn.SiLU(),
            nn.Linear(hidden_dim, height_dim)
        )
        
        self.quantum_l_function = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim),
            nn.SiLU(),
            nn.Linear(hidden_dim, motive_rank)
        )

        # Quantum correction projection
        self.quantum_proj = nn.Linear(2, 2)  # Project within 2D measure space

    def compute_height(self, x: torch.Tensor) -> torch.Tensor:
        """Compute height with quantum corrections.
        
        Args:
            x: Input tensor of shape [batch_size, hidden_dim]
            
        Returns:
            Height tensor of shape [batch_size, height_dim]
        """
        # Classical height
        classical_height = self.height_map(x)
        
        # Quantum correction
        quantum_correction = self.quantum_height(x)
        
        # Combine with quantum weight
        return classical_height + self.quantum_weight * quantum_correction

    def compute_l_function(self, x: torch.Tensor) -> torch.Tensor:
        """Compute L-function with quantum corrections.
        
        Args:
            x: Input tensor of shape [batch_size, hidden_dim]
            
        Returns:
            L-function values of shape [batch_size, motive_rank]
        """
        # Classical L-function
        classical_l = self.l_function(x)
        
        # Quantum correction
        quantum_correction = self.quantum_l_function(x)
        
        # Combine with quantum weight
        return classical_l + self.quantum_weight * quantum_correction

    def compute_dynamics(self, x: torch.Tensor) -> torch.Tensor:
        """Compute arithmetic dynamics with quantum corrections.
        
        Args:
            x: Input tensor of shape [batch_size, hidden_dim] or [batch_size, seq_len, hidden_dim]
            
        Returns:
            Dynamics state tensor with same shape as input
        """
        # Get original shape
        orig_shape = x.shape
        
        # Ensure input is 2D: [batch_size * seq_len, hidden_dim]
        if x.dim() == 3:
            x = x.reshape(-1, self.hidden_dim)
        elif x.dim() == 1:
            x = x.unsqueeze(0)
            
        # Project through L-function network with quantum corrections
        dynamics = self.compute_l_function(x)
        
        # Compute adelic projection
        adelic = self.adelic_proj(x)
        adelic = adelic.view(-1, self.num_primes, self.motive_rank)
        
        # Combine with coupling
        coupled = torch.einsum('bpm,pm->bm', adelic, self.coupling)
        
        # Final dynamics state
        result = dynamics + coupled
        
        # Restore original shape
        if len(orig_shape) == 3:
            result = result.view(orig_shape[0], orig_shape[1], -1)
        
        return result

    def forward(
        self, x: torch.Tensor, steps: int = 1, return_trajectory: bool = False
    ) -> Tuple[torch.Tensor, Dict]:
        """Apply arithmetic dynamics with quantum corrections.

        Args:
            x: Input tensor of shape (batch_size, seq_len, hidden_dim)
            steps: Number of dynamics steps
            return_trajectory: Whether to return full trajectory

        Returns:
            Processed tensor and metrics dictionary
        """
        batch_size, seq_len, _ = x.shape

        # Project to height space with quantum corrections
        x_flat = x.reshape(-1, self.hidden_dim)
        height_coords = self.compute_height(x_flat)
        height_coords = height_coords.view(batch_size, seq_len, self.height_dim)

        # Initialize trajectory storage
        if return_trajectory:
            trajectory = [height_coords.clone()]

        # Apply dynamics
        for _ in range(steps):
            # Compute flow
            flow_field = self.flow(height_coords.reshape(-1, self.height_dim))
            flow_field = flow_field.view(batch_size, seq_len, self.height_dim)

            # Update using exponential map
            height_coords = height_coords + flow_field

            if return_trajectory:
                trajectory.append(height_coords.clone())

        # Compute L-function value with quantum corrections
        l_value = self.compute_l_function(x_flat)
        l_value = l_value.view(batch_size, seq_len, -1)

        # Compute adelic projection
        adelic = self.adelic_proj(x_flat)
        adelic = adelic.view(batch_size, seq_len, self.num_primes, self.motive_rank)

        # Project through measure space and back to input space
        measure = self.measure_proj(height_coords.reshape(-1, self.height_dim))
        output = self.output_proj(measure)
        output = output.view(batch_size, seq_len, self.hidden_dim)

        # Gather metrics
        metrics = {
            "height": self.compute_height(x_flat).mean().item(),
            "l_value": l_value.norm(dim=-1).mean().item(),
            "flow_magnitude": flow_field.norm(dim=-1).mean().item(),
            "adelic_norm": adelic.norm(dim=(-1, -2)).mean().item(),
            "quantum_correction": self.quantum_weight * self.quantum_height(x_flat).norm().item()
        }

        if return_trajectory:
            metrics["trajectory"] = torch.stack(trajectory, dim=1)

        return output, metrics

    def compute_modular_form(self, x: torch.Tensor) -> torch.Tensor:
        """Compute approximate modular form.

        This gives us a way to understand the symmetries in
        our computational patterns through the lens of
        modular forms.
        """
        height_coords = self.height_map(x)

        # Compute modular parameters (simplified)
        q = torch.exp(2 * np.pi * 1j * height_coords[..., 0])

        # Approximate modular form using q-expansion
        powers = torch.arange(1, 5, device=x.device)[None, :]
        q_powers = q[..., None] ** powers

        return q_powers.sum(dim=-1)

    def compute_quantum_correction(self, metric: torch.Tensor) -> torch.Tensor:
        """Compute quantum corrections to the metric.
        
        Args:
            metric: Input metric tensor of shape [batch_size, hidden_dim] or [batch_size, seq_len, hidden_dim]
            
        Returns:
            Quantum correction tensor projected to measure space
        """
        # Get original shape
        orig_shape = metric.shape[:-1]  # Remove last dimension (hidden_dim)
        
        # Ensure input is 2D: [batch_size * seq_len, hidden_dim]
        if metric.dim() == 3:
            metric = metric.reshape(-1, self.hidden_dim)
        elif metric.dim() == 1:
            metric = metric.unsqueeze(0)
        
        # Project metric to height space
        height_coords = self.height_map(metric)
        
        # Compute quantum correction using flow
        correction = self.flow(height_coords)
        
        # Project to measure space (2 dimensions)
        correction = self.measure_proj(correction)  # Project to 2D
        correction = self.quantum_proj(correction)  # Final 2D projection
        
        # Restore original shape if needed
        if len(orig_shape) > 1:
            correction = correction.view(*orig_shape, -1)
        
        return correction

    def compute_quantum_metric(self, x: torch.Tensor) -> torch.Tensor:
        """Compute quantum geometric metric.
        
        Args:
            x: Input tensor
            
        Returns:
            Quantum metric tensor
        """
        # Project to height space
        height_coords = self.height_map(x.reshape(-1, self.hidden_dim))
        height_coords = height_coords.view(*x.shape[:-1], self.height_dim)
        
        # Compute L-function values
        l_values = self.l_function(x.reshape(-1, self.hidden_dim))
        l_values = l_values.view(*x.shape[:-1], self.motive_rank)
        
        # Compute adelic projection
        adelic = self.adelic_proj(x.reshape(-1, self.hidden_dim))
        adelic = adelic.view(*x.shape[:-1], self.num_primes, self.motive_rank)
        
        # Combine components into quantum metric
        quantum_metric = torch.einsum('...i,...j->...ij', height_coords, height_coords)
        l_correction = torch.einsum('...i,...j->...ij', l_values, l_values)
        adelic_correction = torch.einsum('...pi,...pj->...ij', adelic, adelic).mean(dim=-3)
        
        # Project through measure space and back to input space
        combined = quantum_metric + 0.1 * l_correction + 0.01 * adelic_correction
        batch_size = x.shape[0]
        
        # Reshape for projection while preserving batch dimension
        combined_flat = combined.reshape(-1, self.height_dim)  # Flatten all but last dim
        measure = self.measure_proj(combined_flat)  # Project to measure space
        metric = self.output_proj(measure)  # Project back to input space
        
        # Average over height dimensions to get final metric
        metric = metric.reshape(batch_size, self.height_dim, -1).mean(dim=1)
        
        return metric


class ArithmeticPattern(nn.Module):
    """Pattern detection through arithmetic dynamics."""

    def __init__(
        self, input_dim: int, hidden_dim: int, motive_rank: int = 4, num_layers: int = 3
    ):
        super().__init__()

        self.input_dim = input_dim
        self.hidden_dim = hidden_dim
        self.motive_rank = motive_rank
        self.num_layers = num_layers

        # Arithmetic dynamics layers
        self.layers = nn.ModuleList(
            [
                ArithmeticDynamics(
                    hidden_dim=hidden_dim if i > 0 else input_dim,
                    motive_rank=motive_rank,
                )
                for i in range(num_layers)
            ]
        )

        # Pattern projection
        self.pattern_proj = nn.Linear(hidden_dim, hidden_dim)

    def forward(self, x: torch.Tensor) -> Tuple[torch.Tensor, List[Dict]]:
        """
        Apply layered arithmetic dynamics.

        Args:
            x: Input tensor of shape (batch_size, seq_len, hidden_dim)

        Returns:
            - Output tensor of shape (batch_size, seq_len, hidden_dim)
            - List of metrics from each layer
        """
        batch_size, seq_len, hidden_dim = x.shape
        metrics = []

        # Process through layers
        current = x
        for layer in self.layers:
            current, layer_metrics = layer(current)
            metrics.append(
                {
                    "layer_norm": current.norm().item(),
                    "layer_mean": current.mean().item(),
                    "layer_std": current.std().item(),
                    **layer_metrics,  # Include metrics from ArithmeticDynamics
                }
            )

        # Project final output
        output = self.pattern_proj(current)

        return output, metrics


class ModularFormComputer(nn.Module):
    """Compute modular forms for arithmetic dynamics."""
    
    def __init__(
        self,
        hidden_dim: int,
        weight: int = 2,  # Weight of the modular form
        level: int = 1,   # Level of the modular form
        num_coeffs: int = 10,  # Number of q-expansion coefficients
    ):
        super().__init__()
        
        self.hidden_dim = hidden_dim
        self.weight = weight
        self.level = level
        self.num_coeffs = num_coeffs
        
        # Network for computing q-expansion coefficients
        self.coeff_net = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim),
            nn.SiLU(),
            nn.Linear(hidden_dim, num_coeffs * 2)  # Real and imaginary parts
        )
        
        # Network for computing symmetry parameters
        self.symmetry_net = nn.Sequential(
            nn.Linear(hidden_dim, hidden_dim // 2),
            nn.SiLU(),
            nn.Linear(hidden_dim // 2, 2)  # Translation and inversion parameters
        )
        
    def compute_q_expansion(self, x: torch.Tensor) -> torch.Tensor:
        """Compute q-expansion coefficients.
        
        Args:
            x: Input tensor of shape [batch_size, hidden_dim]
            
        Returns:
            Complex coefficients of shape [batch_size, num_coeffs]
        """
        # Get coefficients from network
        coeffs = self.coeff_net(x)
        
        # Split into real and imaginary parts
        real = coeffs[..., :self.num_coeffs]
        imag = coeffs[..., self.num_coeffs:]
        
        # Combine into complex coefficients
        return torch.complex(real, imag)
    
    def compute_symmetries(self, x: torch.Tensor) -> Dict[str, torch.Tensor]:
        """Compute modular symmetry parameters.
        
        Args:
            x: Input tensor of shape [batch_size, hidden_dim]
            
        Returns:
            Dictionary containing symmetry parameters
        """
        params = self.symmetry_net(x)
        
        return {
            'translation': params[..., 0],  # Translation parameter
            'inversion': params[..., 1]     # Inversion parameter
        }
    
    def forward(self, x: torch.Tensor) -> Tuple[torch.Tensor, Dict[str, Any]]:
        """Compute modular form.
        
        Args:
            x: Input tensor of shape [batch_size, hidden_dim]
            
        Returns:
            - q-expansion coefficients
            - Dictionary of symmetry parameters and metrics
        """
        # Compute q-expansion
        q_coeffs = self.compute_q_expansion(x)
        
        # Compute symmetries
        symmetries = self.compute_symmetries(x)
        
        # Compute metrics
        metrics = {
            'weight': self.weight,
            'level': self.level,
            'q_norm': torch.norm(q_coeffs, dim=-1).mean().item(),
            'translation_param': symmetries['translation'].mean().item(),
            'inversion_param': symmetries['inversion'].mean().item()
        }
        
        return q_coeffs, metrics
